refactor: rename *HandlerTest to *TransformerTest
