+++ zeebe/exporter-common/src/test/java/io/camunda/zeebe/exporter/common/auditlog/transformers/AuthorizationAuditLogTransformerTest.java
@@ -11,51 +11,50 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.search.entities.AuditLogEntity.AuditLogOperationType;
 import io.camunda.zeebe.exporter.common.auditlog.AuditLogEntry;
-import io.camunda.zeebe.exporter.common.auditlog.AuditLogInfo;
 import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.ValueType;
-import io.camunda.zeebe.protocol.record.intent.GroupIntent;
-import io.camunda.zeebe.protocol.record.value.GroupRecordValue;
-import io.camunda.zeebe.protocol.record.value.ImmutableGroupRecordValue;
+import io.camunda.zeebe.protocol.record.intent.AuthorizationIntent;
+import io.camunda.zeebe.protocol.record.value.AuthorizationRecordValue;
+import io.camunda.zeebe.protocol.record.value.ImmutableAuthorizationRecordValue;
 import io.camunda.zeebe.test.broker.protocol.ProtocolFactory;
 import java.util.stream.Stream;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class GroupEntityAuditLogHandlerTest {
+class AuthorizationAuditLogTransformerTest {
 
   private final ProtocolFactory factory = new ProtocolFactory();
-  private final GroupEntityAuditLogTransformer transformer = new GroupEntityAuditLogTransformer();
+  private final AuthorizationAuditLogTransformer transformer =
+      new AuthorizationAuditLogTransformer();
 
   public static Stream<Arguments> getIntentMappings() {
     return Stream.of(
-        Arguments.of(GroupIntent.ENTITY_ADDED, AuditLogOperationType.ASSIGN),
-        Arguments.of(GroupIntent.ENTITY_REMOVED, AuditLogOperationType.UNASSIGN));
+        Arguments.of(AuthorizationIntent.CREATED, AuditLogOperationType.CREATE),
+        Arguments.of(AuthorizationIntent.UPDATED, AuditLogOperationType.UPDATE),
+        Arguments.of(AuthorizationIntent.DELETED, AuditLogOperationType.DELETE));
   }
 
   @MethodSource("getIntentMappings")
   @ParameterizedTest
-  void shouldTransformGroupEntityRecord(
-      final GroupIntent intent, final AuditLogOperationType operationType) {
+  void shouldTransformAuthorizationRecord(
+      final AuthorizationIntent intent, final AuditLogOperationType operationType) {
     // given
-    final GroupRecordValue recordValue =
-        ImmutableGroupRecordValue.builder()
-            .from(factory.generateObject(GroupRecordValue.class))
-            .withGroupId("test-group")
-            .withGroupKey(789L)
+    final AuthorizationRecordValue recordValue =
+        ImmutableAuthorizationRecordValue.builder()
+            .from(factory.generateObject(AuthorizationRecordValue.class))
+            .withAuthorizationKey(123L)
             .build();
 
-    final Record<GroupRecordValue> record =
-        factory.generateRecord(ValueType.GROUP, r -> r.withIntent(intent).withValue(recordValue));
+    final Record<AuthorizationRecordValue> record =
+        factory.generateRecord(
+            ValueType.AUTHORIZATION, r -> r.withIntent(intent).withValue(recordValue));
 
     // when
     final var entity = AuditLogEntry.of(record);
     transformer.transform(record, entity);
 
     // then
-    final AuditLogInfo auditLogInfo = AuditLogInfo.of(record);
-    assertThat(auditLogInfo.operationType()).isEqualTo(operationType);
-    assertThat(entity.getEntityKey()).isEqualTo("test-group");
+    assertThat(entity.getOperationType()).isEqualTo(operationType);
   }
 }
