+++ zeebe/exporter-common/src/test/java/io/camunda/zeebe/exporter/common/auditlog/transformers/DecisionAuditLogTransformerTest.java
@@ -22,7 +22,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class DecisionAuditLogHandlerTest {
+class DecisionAuditLogTransformerTest {
 
   private final ProtocolFactory factory = new ProtocolFactory();
   private final DecisionAuditLogTransformer transformer = new DecisionAuditLogTransformer();
@@ -51,7 +51,7 @@ class DecisionAuditLogHandlerTest {
 
     final Record<DecisionRecordValue> record =
         factory.generateRecord(
-            ValueType.DECISION, r -> r.withIntent(DecisionIntent.CREATED).withValue(recordValue));
+            ValueType.DECISION, r -> r.withIntent(intent).withValue(recordValue));
 
     // when
     final var entity = AuditLogEntry.of(record);
@@ -63,5 +63,6 @@ class DecisionAuditLogHandlerTest {
     assertThat(entity.getDecisionDefinitionKey()).isEqualTo(789L);
     assertThat(entity.getDecisionRequirementsId()).isEqualTo("decisionRequirementsId");
     assertThat(entity.getDecisionDefinitionId()).isEqualTo("decisionId");
+    assertThat(entity.getOperationType()).isEqualTo(operationType);
   }
 }
