+++ zeebe/exporter-common/src/test/java/io/camunda/zeebe/exporter/common/auditlog/transformers/GroupEntityAuditLogTransformerTest.java
@@ -11,7 +11,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.search.entities.AuditLogEntity.AuditLogOperationType;
 import io.camunda.zeebe.exporter.common.auditlog.AuditLogEntry;
-import io.camunda.zeebe.exporter.common.auditlog.AuditLogInfo;
 import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.protocol.record.intent.GroupIntent;
@@ -23,21 +22,20 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class GroupAuditLogHandlerTest {
+class GroupEntityAuditLogTransformerTest {
 
   private final ProtocolFactory factory = new ProtocolFactory();
-  private final GroupAuditLogTransformer transformer = new GroupAuditLogTransformer();
+  private final GroupEntityAuditLogTransformer transformer = new GroupEntityAuditLogTransformer();
 
   public static Stream<Arguments> getIntentMappings() {
     return Stream.of(
-        Arguments.of(GroupIntent.CREATED, AuditLogOperationType.CREATE),
-        Arguments.of(GroupIntent.UPDATED, AuditLogOperationType.UPDATE),
-        Arguments.of(GroupIntent.DELETED, AuditLogOperationType.DELETE));
+        Arguments.of(GroupIntent.ENTITY_ADDED, AuditLogOperationType.ASSIGN),
+        Arguments.of(GroupIntent.ENTITY_REMOVED, AuditLogOperationType.UNASSIGN));
   }
 
   @MethodSource("getIntentMappings")
   @ParameterizedTest
-  void shouldTransformGroupRecord(
+  void shouldTransformGroupEntityRecord(
       final GroupIntent intent, final AuditLogOperationType operationType) {
     // given
     final GroupRecordValue recordValue =
@@ -55,8 +53,7 @@ class GroupAuditLogHandlerTest {
     transformer.transform(record, entity);
 
     // then
-    final AuditLogInfo auditLogInfo = AuditLogInfo.of(record);
-    assertThat(auditLogInfo.operationType()).isEqualTo(operationType);
     assertThat(entity.getEntityKey()).isEqualTo("test-group");
+    assertThat(entity.getOperationType()).isEqualTo(operationType);
   }
 }
