+++ zeebe/exporter-common/src/test/java/io/camunda/zeebe/exporter/common/auditlog/transformers/TenantEntityAuditLogTransformerTest.java
@@ -11,7 +11,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.search.entities.AuditLogEntity.AuditLogOperationType;
 import io.camunda.zeebe.exporter.common.auditlog.AuditLogEntry;
-import io.camunda.zeebe.exporter.common.auditlog.AuditLogInfo;
 import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.protocol.record.intent.TenantIntent;
@@ -23,21 +22,20 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class TenantAuditLogHandlerTest {
+class TenantEntityAuditLogTransformerTest {
 
   private final ProtocolFactory factory = new ProtocolFactory();
-  private final TenantAuditLogTransformer transformer = new TenantAuditLogTransformer();
+  private final TenantEntityAuditLogTransformer transformer = new TenantEntityAuditLogTransformer();
 
   public static Stream<Arguments> getIntentMappings() {
     return Stream.of(
-        Arguments.of(TenantIntent.CREATED, AuditLogOperationType.CREATE),
-        Arguments.of(TenantIntent.UPDATED, AuditLogOperationType.UPDATE),
-        Arguments.of(TenantIntent.DELETED, AuditLogOperationType.DELETE));
+        Arguments.of(TenantIntent.ENTITY_ADDED, AuditLogOperationType.ASSIGN),
+        Arguments.of(TenantIntent.ENTITY_REMOVED, AuditLogOperationType.UNASSIGN));
   }
 
   @MethodSource("getIntentMappings")
   @ParameterizedTest
-  void shouldTransformTenantRecord(
+  void shouldTransformTenantEntityRecord(
       final TenantIntent intent, final AuditLogOperationType operationType) {
     // given
     final TenantRecordValue recordValue =
@@ -55,8 +53,7 @@ class TenantAuditLogHandlerTest {
     transformer.transform(record, entity);
 
     // then
-    final AuditLogInfo auditLogInfo = AuditLogInfo.of(record);
-    assertThat(auditLogInfo.operationType()).isEqualTo(operationType);
     assertThat(entity.getEntityKey()).isEqualTo("test-tenant");
+    assertThat(entity.getOperationType()).isEqualTo(operationType);
   }
 }
